64 Bit Warnings fix

+++ UIImage+PDF/PDFView.h
@@ -21,7 +21,7 @@
 +(CGRect) mediaRectForURL:(NSURL *)resourceURL;
 +(CGRect) mediaRectForURL:(NSURL *)resourceURL atPage:(int)page;
 +(CGRect) mediaRectForData:(NSData *)data atPage:(int)page;
-+(int) pageCountForURL:(NSURL *)resourceURL;
++(NSInteger) pageCountForURL:(NSURL *)resourceURL;
 +(NSURL *)resourceURLForName:(NSString *)resourceName;
 +(void)renderIntoContext:(CGContextRef)ctx url:(NSURL *)resourceURL data:(NSData *)resourceData size:(CGSize)size page:(int)page;
